modified:   PKGBUILD 	modified:   build.sh 	modified:   config.h

+++ build.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 
 HASH='61bb8b2241d4db08bea4261c82e27cd9797099e7'
+
 start_dir="${PWD}"
-build_dir="/tmp/sl_make"
+build_dir="/tmp/suckless_dwm"
+
+deps=(sudo mpd st dmenu git libx11 libxinerama libxft freetype2)
+for dep in ${deps[@]}; do
+	msg=$(pacman -Q $dep 2>&1 1>/dev/null)
+	msg=${msg#*\'}
+	msg=${msg%\'*}
+	depserr+=($msg)
+	unset msg
+done
+
+if [[ -n ${depserr[@]} ]]; then
+	echo "No dependencies: ${depserr[@]}"
+	exit 1
+fi
 
 [[ -d "${build_dir}" ]] && rm -rf "${build_dir}"
